Moved config file to soepy/soepy.
