+++ development/tests/property/property_auxiliary.py
@@ -9,7 +9,7 @@ import os
 
 import numpy as np
 
-from soepy.python.soepy_config import PACKAGE_DIR
+from soepy.soepy_config import PACKAGE_DIR
 
 
 def collect_tests():
