+++ development/tests/regression/run.py
@@ -7,7 +7,7 @@ import argparse
 import numpy as np
 
 from soepy.python.simulate.simulate_python import simulate
-from soepy.python.soepy_config import TEST_RESOURCES_DIR
+from soepy.soepy_config import TEST_RESOURCES_DIR
 from soepy.test.random_init import random_init
 from soepy.test.random_init import print_dict
 from soepy.test.auxiliary import cleanup
