+++ soepy/__init__.py
@@ -4,8 +4,8 @@ import os
 import pytest
 
 from soepy.python.simulate.simulate_python import simulate
-from soepy.python.soepy_config import PACKAGE_DIR
-import soepy.python.soepy_config
+from soepy.soepy_config import PACKAGE_DIR
+import soepy.soepy_config
 
 
 def test():
