+++ soepy/soepy_config.py
@@ -8,6 +8,5 @@ import numpy as np
 np.testing.assert_equal(sys.version_info[:2] >= (3, 6), True)
 
 # We rely on relative paths throughout the package.
-ROOT_DIR = Path(__file__).parent.parent.absolute()
-PACKAGE_DIR = ROOT_DIR.parents[0]
+PACKAGE_DIR = Path(__file__).parent.parent.absolute()
 TEST_RESOURCES_DIR = PACKAGE_DIR / "soepy" / "test" / "resources"
